patient file endpoint
