+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/NBS6PatientFileRedirection.java
@@ -1,6 +1,5 @@
 package gov.cdc.nbs.patient;
 
-import gov.cdc.nbs.patient.profile.redirect.outgoing.NBS6PatientFileRedirectResolver;
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,11 +8,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 @Hidden
 @RestController
-class NBS6PatientRedirection {
+class NBS6PatientFileRedirection {
 
   private final NBS6PatientFileRedirectResolver resolver;
 
-  NBS6PatientRedirection(final NBS6PatientFileRedirectResolver resolver) {
+  NBS6PatientFileRedirection(final NBS6PatientFileRedirectResolver resolver) {
     this.resolver = resolver;
   }
