+++ apps/modernization-ui/src/apps/search/patient/result/patientSearchResult.tsx
@@ -3,6 +3,7 @@ import { displayNameElement, matchesLegalName } from 'name';
 import { displayAddress } from 'address/display';
 
 import { ItemGroup } from 'design-system/item';
+import { internalizeDate, displayAgeAsOfToday } from 'date';
 
 // Displays Other names, that are not the legal name
 const displayOtherNames = (result: PatientSearchResult, order: 'normal' | 'reverse' = 'normal'): JSX.Element => {
