+++ apps/modernization-ui/src/date/index.ts
@@ -4,4 +4,4 @@ export { internalizeDate } from './InternalizeDate';
 export { calculateAge } from './calculateAge';
 export { today } from './today';
 export * from './asLocalDate';
-export * from './calculateAge';
+export { displayAgeAsOfToday } from './displayAge';
